requirements.txt
# Context Clock Requirements
# Core dependencies for the Context Clock Windows application

# GUI Framework
PyQt5>=5.15.0

# Configuration Management
PyYAML>=6.0

# System Process Management
psutil>=5.9.0

# Audio Playback (Primary)
pygame>=2.5.0

# Audio Playback (Fallback - optional due to compatibility issues)
# playsound==1.3.0  # Commented out due to Python 3.12+ compatibility issues

# Task Scheduling (Optional - for future enhancements)
APScheduler>=3.10.0

# Windows API Access (Already included in Python on Windows)
# pywin32>=306  # Uncomment if needed for additional Windows features

# Development and Building
# pyinstaller>=6.0.0  # Uncomment when building .exe

# Optional Dependencies for Enhanced Features
# requests>=2.31.0  # For potential web-based configuration
# Pillow>=10.0.0    # For advanced image processing if needed 